pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "logging>=0.4.9.6",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
]
